+++ app/assets/javascripts/app/collections/aspects.js
@@ -1,7 +1,7 @@
 // @license magnet:?xt=urn:btih:0b31508aeb0634b347b8270c7bee4d411b5d4109&dn=agpl-3.0.txt AGPL-v3-or-Later
 
 app.collections.Aspects = Backbone.Collection.extend({
-  model: app.models.Aspect,
+  model: app.models.AspectSelection,
 
   selectedAspects: function(attribute){
     return _.pluck(_.filter(this.toJSON(), function(a){