+++ app/assets/javascripts/app/views/aspects_list_view.js
@@ -6,18 +6,26 @@ app.views.AspectsList = app.views.Base.extend({
   el: '#aspects_list',
 
   events: {
-    'click .toggle_selector' : 'toggleAll'
+    "click .toggle_selector" : "toggleAll"
+  },
+
+  subviews : {
+    "#newAspectContainer" : "aspectCreateView"
   },
 
   initialize: function() {
-    this.collection.on('change', this.toggleSelector, this);
-    this.collection.on('change', this.updateStreamTitle, this);
-    this.collection.on('aspectStreamFetched', this.updateAspectList, this);
+    this.collection.on("change", this.toggleSelector, this);
+    this.collection.on("change", this.updateStreamTitle, this);
+    this.collection.on("aspectStreamFetched", this.updateAspectList, this);
+    app.events.on("aspect:create", function(id) { window.location = "/contacts?a_id=" + id });
+  },
+
+  aspectCreateView: function() {
+    return new app.views.AspectCreate();
   },
 
   postRenderTemplate: function() {
     this.collection.each(this.appendAspect, this);
-    this.$('a[rel*=facebox]').facebox();
     this.updateStreamTitle();
     this.toggleSelector();
   },
@@ -67,6 +75,6 @@ app.views.AspectsList = app.views.Base.extend({
 
   hideAspectsList: function() {
     this.$el.empty();
-  },
+  }
 });
 // @license-end